// wardn/frontend/lib/api/generated/model/resourceLimitUpsert.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.0.1
 */
import type { ResourceLimitUpsertScopeType } from './resourceLimitUpsertScopeType';

export interface ResourceLimitUpsert {
  /**
     * @minLength 1
     * @maxLength 120
     */
  limitKey: string;
  scopeId?: string | null;
  scopeType: ResourceLimitUpsertScopeType;
  /** @minimum 0 */
  value: number;
}